+++ src/controllers/city.controller.js
@@ -1,12 +1,24 @@
 
+import { decodeAccessToken } from '../middleware/token.js';
 import { City } from '../models/city.model.js'; // Importa el modelo Company que defines en otro archivo
-import {getAllModels, getModelById, createModel, updateModel, deleteModel, getModelByParameterOne, getModelByParameterMany} from "./general.controller.js"
+import {getAllModels, getModelById, createModel, updateModel, deleteModel, getModelByParameterOne, getModelByParameterMany, hasPermissRol, hasPermissUser, searchOperation, addOperation, updateOperation, deleteOperation} from "./general.controller.js"
 
 
+const module = "City"
 
 //Metodo que devuelve todos las ciudades
 export const getCities = async(req,res)=> {
         const result = await getAllModels(City);
+        const token = req.headers.authorization;   
+        
+        if(!token) return res.status(401).json({message:"Token is required"})
+        const dataToken = decodeAccessToken(token);
+     
+        const rolCanGet = await hasPermissRol(dataToken, searchOperation, module)
+        const userCanGet = await hasPermissUser(dataToken, searchOperation, module)
+    
+        if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
         if (result.success) {
             res.status(result.status).json(result.models);
         } else {
@@ -19,6 +31,16 @@ export const getCities = async(req,res)=> {
 export const getCityById = async(req,res)=>{
     const { id } = req.params;
     const result = await getModelById(City, id);
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, searchOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, searchOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
     if (result.success) {
         res.status(result.status).json(result.model);
     } else {
@@ -31,7 +53,19 @@ export const getCityById = async(req,res)=>{
 //Parametros: name, id_department
 export const createCity =  async(req,res)=> {
     const { name, id_department, postal_code } = req.body;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, addOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, addOperation, module)
 
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
+    if(!name || !id_department || !postal_code){
+        return res.status(400).json({ message: 'Please fill all fields' });
+    }
 
     const nameLower = name.toLowerCase();
     const nameCapitalize = nameLower.charAt(0).toUpperCase() + nameLower.slice(1);
@@ -41,7 +75,7 @@ export const createCity =  async(req,res)=> {
         return res.status(400).json({ message: 'Cannot create a duplicate city' });
     }
 
-    const result = await createModel(City, { nameCapitalize, id_department, postal_code });
+    const result = await createModel(City, { name:nameCapitalize, id_department, postal_code });
     if (result.success) {
         res.status(result.status).json({ message: 'City created' });
     } else {
@@ -55,7 +89,17 @@ export const createCity =  async(req,res)=> {
 //Parametros: name, id_department
 export const updateCity = async(req,res)=>{
     const { id } = req.params;
-    const { name, id_department } = req.body;
+    let { name, id_department } = req.body;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, updateOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, updateOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
 
     const city = await getModelById(City,id)
     const nameLower = name.toLowerCase();
@@ -87,6 +131,16 @@ export const updateCity = async(req,res)=>{
 //Parametros: id
 export const deleteCity = async(req,res)=>{
     const { id } = req.params;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, deleteOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, deleteOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
     const result = await deleteModel(City, id);
     if (result.success) {
         res.status(result.status).json({ message: 'City deleted' });
