+++ src/controllers/company.controller.js
@@ -1,13 +1,24 @@
 
+import { decodeAccessToken } from '../middleware/token.js';
 import { Company } from '../models/company.model.js'; // Importa el modelo Company que defines en otro archivo
-import {getAllModels, getModelById, createModel, updateModel, deleteModel, getModelByParameterMany, getModelByParameterOne} from "./general.controller.js"
+import {getAllModels, getModelById, createModel, updateModel, deleteModel, getModelByParameterMany, getModelByParameterOne, searchOperation, addOperation, updateOperation, deleteOperation, hasPermissRol, hasPermissUser} from "./general.controller.js"
 
 const namePrimaryKey='nit'
-
+const module = "Company"
 
 //Metodo que devuelve todos las compañias
 export const getCompanies = async(req,res)=> {
         const result = await getAllModels(Company);
+        const token = req.headers.authorization;   
+        
+        if(!token) return res.status(401).json({message:"Token is required"})
+        const dataToken = decodeAccessToken(token);
+     
+        const rolCanGet = await hasPermissRol(dataToken, searchOperation, module)
+        const userCanGet = await hasPermissUser(dataToken, searchOperation, module)
+    
+        if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
         if (result.success) {
             res.status(result.status).json(result.models);
         } else {
@@ -19,6 +30,16 @@ export const getCompanies = async(req,res)=> {
 //Parametros: nit
 export const getCompanyById = async(req,res)=>{
     const { id } = req.params;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, searchOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, searchOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
     const result = await getModelById(Company, id);
     if (result.success) {
         res.status(result.status).json(result.model);
@@ -32,6 +53,16 @@ export const getCompanyById = async(req,res)=>{
 //Parametros: nit, name, id_masteruser
 export const createCompany =  async(req,res)=> {
     const { nit, name, photo, address, phone, status, id_city } = req.body;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, addOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, addOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
 
     const nameLower = name.toLowerCase();
     const nameCapitalize = nameLower.charAt(0).toUpperCase() + nameLower.slice(1);
@@ -42,7 +73,7 @@ export const createCompany =  async(req,res)=> {
     }
 
 
-    const result = await createModel(Company, { nit, nameCapitalize, photo, address, phone, status, id_city });
+    const result = await createModel(Company, { nit, name:nameCapitalize, photo, address, phone, status, id_city });
 
 
 
@@ -60,6 +91,15 @@ export const createCompany =  async(req,res)=> {
 export const updateCompany = async(req,res)=>{
     const { nit } = req.params;
     let { name, photo,phone, status, id_city } = req.body;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, updateOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, updateOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
 
 
     const nameLower = name.toLowerCase();
@@ -98,6 +138,16 @@ export const updateCompany = async(req,res)=>{
 //Parametros: id
 export const deleteCompany = async(req,res)=>{
     const { id } = req.params;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, deleteOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, deleteOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
     const result = await deleteModel(Company, id, namePrimaryKey);
     if (result.success) {
         res.status(result.status).json({ message: 'Company deleted' });
@@ -111,6 +161,16 @@ export const deleteCompany = async(req,res)=>{
 //Parametros: nit
 export const getCompanyOnlyActive = async(req,res)=>{
     const { id } = req.params;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, searchOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, searchOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
     const result = await getModelByParameterMany(Company, "status", true);
     if (result.success) {
         res.status(result.status).json(result.model);
@@ -123,6 +183,16 @@ export const getCompanyOnlyActive = async(req,res)=>{
 //Parametros: nit
 export const getCompanyOnlyInactive = async(req,res)=>{
     const { id } = req.params;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, searchOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, searchOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
     let result = await getModelByParameterMany(Company, "status", false);
     if (result.success) {
         res.status(result.status).json(result.model);
