+++ src/controllers/country.controller.js
@@ -1,13 +1,24 @@
 
+import { decodeAccessToken } from '../middleware/token.js';
 import { Country } from '../models/country.model.js'; // Importa el modelo Company que defines en otro archivo
-import {getAllModels, getModelById, createModel, updateModel, deleteModel, getModelByParameterOne, getModelByParameterMany} from "./general.controller.js"
+import {getAllModels, getModelById, createModel, updateModel, deleteModel, getModelByParameterOne, getModelByParameterMany, hasPermissRol, hasPermissUser, searchOperation, addOperation, updateOperation, deleteOperation} from "./general.controller.js"
 
 
-
+const module = "Country"
 
 //Metodo que devuelve todos los paises
 export const getCountries = async(req,res)=> {
         const result = await getAllModels(Country);
+        const token = req.headers.authorization;   
+        
+        if(!token) return res.status(401).json({message:"Token is required"})
+        const dataToken = decodeAccessToken(token);
+     
+        const rolCanGet = await hasPermissRol(dataToken, searchOperation, module)
+        const userCanGet = await hasPermissUser(dataToken, searchOperation, module)
+    
+        if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
         if (result.success) {
             res.status(result.status).json(result.models);
         } else {
@@ -19,6 +30,16 @@ export const getCountries = async(req,res)=> {
 //Parametros: id
 export const getCountryById = async(req,res)=>{
     const { id } = req.params;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, searchOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, searchOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
     const result = await getModelById(Country, id);
     if (result.success) {
         res.status(result.status).json(result.model);
@@ -32,6 +53,16 @@ export const getCountryById = async(req,res)=>{
 //Parametros:  name
 export const createCountry =  async(req,res)=> {
     const {name } = req.body;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, addOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, addOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
 
     const nameLower = name.toLowerCase();
     const nameCapitalize = nameLower.charAt(0).toUpperCase() + nameLower.slice(1);
@@ -40,7 +71,7 @@ export const createCountry =  async(req,res)=> {
         return res.status(400).json({ message: 'Cannot create a duplicate country' });
     }
 
-    const result = await createModel(Country, { nameCapitalize });
+    const result = await createModel(Country, { name:nameCapitalize });
     if (result.success) {
         res.status(result.status).json({ message: 'Country created' });
     } else {
@@ -54,7 +85,17 @@ export const createCountry =  async(req,res)=> {
 //Parametros: id, name
 export const updateCountry = async(req,res)=>{
     const { id } = req.params;
-    const { name } = req.body;
+    let { name } = req.body;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, updateOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, updateOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
 
     const country = await getModelById(Country,id)
 
@@ -87,6 +128,16 @@ export const updateCountry = async(req,res)=>{
 //Parametros: id
 export const deleteCountry = async(req,res)=>{
     const { id } = req.params;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, deleteOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, deleteOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
     const result = await deleteModel(Country, id);
     if (result.success) {
         res.status(result.status).json({ message: 'Country deleted' });
@@ -100,6 +151,16 @@ export const deleteCountry = async(req,res)=>{
 //Parametros: name
 export const getCountryByName = async(req,res)=>{
     const { name } = req.params;
+    const token = req.headers.authorization;   
+        
+    if(!token) return res.status(401).json({message:"Token is required"})
+    const dataToken = decodeAccessToken(token);
+ 
+    const rolCanGet = await hasPermissRol(dataToken, searchOperation, module)
+    const userCanGet = await hasPermissUser(dataToken, searchOperation, module)
+
+    if(!rolCanGet && userCanGet ){ return res.status(403).json({ message: 'User not has necessary permissions ' }); }
+
     const nameLower = name.toLowerCase();
     const nameCapitalize = nameLower.charAt(0).toUpperCase() + nameLower.slice(1);
